+++ src/sophronia/app/country/[slug]/getCountry.js
@@ -1,0 +1,7 @@
+import { getBackendURL } from 'lib/ApiFetcher';
+
+export async function getCountry(slug) {
+	const url = getBackendURL();
+	const res = await fetch(`${url}country/${slug}`, { cache: 'no-store' });
+	return res.json();
+}
