+++ src/sophronia/app/distributor/[slug]/head.tsx
@@ -1,4 +1,4 @@
-import { getDistributor } from './page';
+import { getDistributor } from './getDistributor';
 
 export default async function Head({ params }: { params: { slug: string } }) {
 	const distributor = await getDistributor(params.slug);
