+++ src/sophronia/app/film/[slug]/getFilm.js
@@ -1,0 +1,7 @@
+import { getBackendURL } from 'lib/ApiFetcher';
+
+export async function getFilm(slug) {
+	const url = getBackendURL();
+	const res = await fetch(`${url}film/${slug}`, { cache: 'no-store' });
+	return res.json();
+}
